# Add performance tests for influx serialization