--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -38,44 +38,55 @@ class Serializer {
     toInflux(point, nameField) {
         if (!nameField) { throw new Error('nameField argument required'); }
 
+        // Measurement name
         var name = point[nameField];
-        var timestamp_ns = point.time ? point.time.unixms() : null;
-
-        point = _.omit(point, nameField, 'time');
-
-        var tags = _.pick(point, this._isTag.bind(this));
-        var vals = _.omit(point, this._isTag.bind(this));
-
-        if (_.keys(vals).length === 0) { throw new Error('point requires at least one field'); }
         if (!name) { throw new Error('point is missing a name'); }
 
-        var key = _.flatten([name, _.map(tags, this._concatKeyVal)]).map(this._escapeTag);
+        // Timestamp
+        var timestamp = point.time ? point.time.unixms() : '';
+        var point = _.omit(point, nameField, 'time');
 
-        var val = _.chain(vals)
-                   .map(this._serializeValue.bind(this))
-                   .object()
-                   .map(this._concatKeyVal)
-                   .value();
+        // Tags, values
+        var tags = _.pick(point, (val, key) => {
+            return _.isString(val) && !_.contains(this.valFields, key);
+        });
 
-        return _.compact([key.join(","), val.join(","), timestamp_ns]).join(" ");
+        var vals = _.omit(point, (val, key) => {
+            return _.isString(val) && !_.contains(this.valFields, key);
+        });
+
+        // Drop point early, influx wont accept this
+        if (_.keys(vals).length === 0) { throw new Error('point requires at least one field'); }
+
+        // left column - name + tags
+        var taggedName = [
+            this._escapeTag(name)
+        ].concat(
+            _.map(tags, (val, key) => {
+                return this._escapeTag(`${key}=${val}`);
+            })
+        ).join(",");
+
+        // middle column
+        var values = _.map(vals, (val, key) => {
+            return `${key}=${this._serializeValue(key, val, this.intFields)}`;
+        }).join(",");
+
+        return `${taggedName} ${values} ${timestamp}`.trim();
     }
 
-    _concatKeyVal(v, k) {
-        return `${k}=${v}`;
-    }
-
-    _serializeValue(v, k) {
+    _serializeValue(k, v, intFields) {
         if (values.isBoolean(v)) {
-            return [k, (v ? 't' : 'f')];
+            return (v ? 't' : 'f');
         } else
         if (values.isString(v)) {
-            return [k, `"${this._escapeString(v)}"`];
+            return `"${this._escapeString(v)}"`;
         } else
-        if (this._isInt(v, k)) {
-            return [k, `${Math.floor(v)}i`];
+        if (_.isNumber(v) && _.contains(intFields, k)) {
+            return `${Math.floor(v)}i`;
         } else
         if (values.isDate(v) || values.isDuration(v)) {
-            return [k, v.unixms()];
+            return v.unixms();
         } else
         if (values.isRegExp(v)) {
             throw new Error('Serializing regular expression values is not supported.');
@@ -86,27 +97,18 @@ class Serializer {
         if (values.isObject(v) || values.isArray(v)) {
             throw new Error('Serializing array and object fields is not supported.');
         } else {
-            return [k, v];
+            return v;
         }
     }
 
     // https://influxdb.com/docs/v0.9/write_protocols/line.html#key
     _escapeTag(str) {
-        var specialChars = /[\s,]/g;
-        return str.replace(specialChars, m => { return `\\${m}`; });
+        return str.replace(/[\s,]/g, "\\$&");
     }
 
     // https://influxdb.com/docs/v0.9/write_protocols/line.html#fields (Strings)
     _escapeString(str) {
         return str.replace(/"/g, '\\"');
-    }
-
-    _isTag(val, key, obj) {
-        return _.isString(val) && !_.contains(this.valFields, key);
-    }
-
-    _isInt(val, key, obj) {
-        return _.isNumber(val) && _.contains(this.intFields, key);
     }
 }