--- a/lib/write.js
+++ b/lib/write.js
@@ -61,10 +61,14 @@ class WriteInflux extends AdapterWrite {
 
         var chunks = this.writeChunks(body, this.writeChunkSize);
 
+        return this.sendWrites(reqUrl, chunks);
+    }
+
+    sendWrites(url, chunks) {
         this.pendingWrites++;
 
         return Promise.each(chunks, (chunk) => {
-            return this.request(reqUrl, chunk);
+            return this.request(url, chunk);
         }).catch((err) => {
             this.trigger('error', err);
         }).then(() => {
@@ -81,9 +85,9 @@ class WriteInflux extends AdapterWrite {
         return writeChunks;
     }
 
-    request(reqUrl, chunk) {
+    request(url, chunk) {
         return request.async({
-            url: reqUrl,
+            url: url,
             method: 'POST',
             body: chunk.join('\n')
         }).then((response) => {