--- a/lib/write.js
+++ b/lib/write.js
@@ -67,22 +67,26 @@ class WriteInflux extends AdapterWrite {
         this.pendingWrites++;
 
         return Promise.each(writeChunks, (chunk) => {
-            return request.async({
-                url: reqUrl,
-                method: 'POST',
-                body: chunk.join('\n')
-            }).then((response) => {
-                // https://influxdb.com/docs/v0.9/guides/writing_data.html#writing-data-using-the-http-api
-                // section http response summary
-                if (response.statusCode === 200 || response.statusCode > 300) {
-                    throw this.runtimeError('INTERNAL-ERROR', { error: response.body });
-                }
-            });
+            return this.request(reqUrl, chunk);
         }).catch((err) => {
             this.trigger('error', err);
         }).then(() => {
             this.pendingWrites--;
             this.checkIfDone();
+        });
+    }
+
+    request(reqUrl, chunk) {
+        return request.async({
+            url: reqUrl,
+            method: 'POST',
+            body: chunk.join('\n')
+        }).then((response) => {
+            // https://influxdb.com/docs/v0.9/guides/writing_data.html#writing-data-using-the-http-api
+            // section http response summary
+            if (response.statusCode === 200 || response.statusCode > 300) {
+                throw this.runtimeError('INTERNAL-ERROR', { error: response.body });
+            }
         });
     }