--- a/lib/write.js
+++ b/lib/write.js
@@ -59,14 +59,11 @@ class WriteInflux extends AdapterWrite {
             }
         }));
 
-        var writeChunks = [];
-        while (body.length > 0) {
-            writeChunks.push(body.splice(0, this.writeChunkSize));
-        }
+        var chunks = this.writeChunks(body, this.writeChunkSize);
 
         this.pendingWrites++;
 
-        return Promise.each(writeChunks, (chunk) => {
+        return Promise.each(chunks, (chunk) => {
             return this.request(reqUrl, chunk);
         }).catch((err) => {
             this.trigger('error', err);
@@ -74,6 +71,14 @@ class WriteInflux extends AdapterWrite {
             this.pendingWrites--;
             this.checkIfDone();
         });
+    }
+
+    writeChunks(body, writeChunkSize) {
+        var writeChunks = [];
+        while (body.length > 0) {
+            writeChunks.push(body.splice(0, writeChunkSize));
+        }
+        return writeChunks;
     }
 
     request(reqUrl, chunk) {