--- a/lib/write.js
+++ b/lib/write.js
@@ -48,7 +48,15 @@ class WriteInflux extends AdapterWrite {
 
         reqUrl = url.format(parsedUrl);
 
-        var body = _.compact(_.map(points, (p) => {
+        var rows = this.serialize(points);
+
+        var chunks = this.chunkRows(rows, this.writeChunkSize);
+
+        return this.sendWrites(reqUrl, chunks);
+    }
+
+    serialize(points) {
+        return _.compact(_.map(points, (p) => {
             try {
                 return this.serializer.toInflux(p, this.nameField);
             } catch(err) {
@@ -58,10 +66,6 @@ class WriteInflux extends AdapterWrite {
                 return null;
             }
         }));
-
-        var chunks = this.writeChunks(body, this.writeChunkSize);
-
-        return this.sendWrites(reqUrl, chunks);
     }
 
     sendWrites(url, chunks) {
@@ -77,7 +81,7 @@ class WriteInflux extends AdapterWrite {
         });
     }
 
-    writeChunks(body, writeChunkSize) {
+    chunkRows(body, writeChunkSize) {
         var writeChunks = [];
         while (body.length > 0) {
             writeChunks.push(body.splice(0, writeChunkSize));