--- a/package.json
+++ b/package.json
@@ -15,7 +15,7 @@
   "scripts": {
     "test": "mocha test/**/*.spec.js test/*.spec.js",
     "lint": "eslint --ignore-pattern=./coverage/ .",
-    "cov": "istanbul cover _mocha && istanbul check-coverage",
+    "cov": "istanbul cover _mocha test/**/*.spec.js test/*.spec.js && istanbul check-coverage",
     "perf": "node ./test/performance/index.js"
   },
   "dependencies": {