--- a/package.json
+++ b/package.json
@@ -13,7 +13,7 @@
   "main": "index.js",
   "repository": "juttle/juttle-influx-adapter",
   "scripts": {
-    "test": "mocha",
+    "test": "mocha test/**/*.spec.js test/*.spec.js",
     "lint": "eslint --ignore-pattern=./coverage/ .",
     "cov": "istanbul cover _mocha && istanbul check-coverage",
     "perf": "node ./test/performance/index.js"