--- a/test/influx-integration.spec.js
+++ b/test/influx-integration.spec.js
@@ -8,9 +8,6 @@ var path = require('path');
 var Promise = require('bluebird');
 var retry = require('bluebird-retry');
 
-var request = Promise.promisifyAll(require('request'));
-request.async = Promise.promisify(request);
-
 var juttle_test_utils = require('juttle/test').utils;
 var check_juttle = juttle_test_utils.check_juttle;
 
@@ -19,82 +16,22 @@ var retry_options = {
     timeout: 2000
 };
 
-var influx_api_url = {
+var influx_api_url = url.format({
     protocol: 'http',
     hostname: process.env.INFLUX_HOST || 'localhost',
     port: process.env.INFLUX_PORT || 8086,
     pathname: '/'
-};
+});
+
+var DB = require('./test_utils').DB.init(influx_api_url);
 
 juttle_test_utils.configureAdapter({
     influx: {
         path: path.resolve(__dirname, '..'),
-        url: url.format(influx_api_url)
+        url: influx_api_url
     }
 });
 
-/* DB utils */
-var DB = {
-    // Influx doesnt seem to like writing future points?
-    _t0: Date.now() - 3600 * 1000,
-    _points: 10,
-    _dt: 1000,
-
-    _handle_response(response) {
-        if (response.statusCode !== 200 && response.statusCode !== 204) {
-            throw new Error(['error', response.statusCode, response.body].join(' '));
-        }
-
-        return response.body === "" ? null : JSON.parse(response.body);
-    },
-
-    _fixture() {
-        var payload = "";
-
-        for (var i = 0; i < this._points; i++) {
-            var t_i = this._t0 + i * this._dt;
-            payload += `cpu,host=host${i} value=${i} ${t_i}\n`;
-        }
-
-        for (var j = 0; j < this._points; j++) {
-            var t_j = this._t0 + j * this._dt;
-            payload += `mem,host=host${j} value=${j} ${t_j}\n`;
-        }
-
-        return payload;
-    },
-
-    query(q) {
-        var requestUrl = _.extend(influx_api_url, { pathname: '/query', query: { q, 'db': 'test' } });
-        return request.async({
-            url: url.format(requestUrl),
-            method: 'GET'
-        }).then(this._handle_response).catch((e) => {
-            throw e;
-        });
-    },
-
-    create() {
-        return this.query('CREATE DATABASE test');
-    },
-
-    drop() {
-        return this.query('DROP DATABASE test');
-    },
-
-    insert(data) {
-        var payload = data || this._fixture();
-        var requestUrl = _.extend(influx_api_url, { pathname: '/write', query: { 'db': 'test', 'precision': 'ms' } });
-
-        return request.async({
-            url: url.format(requestUrl),
-            method: 'POST',
-            body: payload
-        }).then(this._handle_response).catch((e) => {
-            throw e;
-        });
-    },
-};
 
 describe('@integration influxdb tests', () => {
     describe('read', () => {