--- a/test/performance/index.js
+++ b/test/performance/index.js
@@ -1,0 +1,2 @@
+require('./serializer/influx');
+require('./serializer/juttle');