--- a/test/performance/index.js
+++ b/test/performance/index.js
@@ -1,2 +1,5 @@
 require('./serializer/influx');
 require('./serializer/juttle');
+require('./write/serialize.js');
+require('./write/stubbed.js');
+require('./write/live.js');