--- a/test/test_utils.js
+++ b/test/test_utils.js
@@ -1,5 +1,0 @@
-var parseFilter = require('./test_utils/parse_filter');
-
-module.exports = {
-    parseFilter
-};