--- a/test/test_utils/index.js
+++ b/test/test_utils/index.js
@@ -1,5 +1,7 @@
 var parseFilter = require('./parse_filter');
+var DB = require('./db');
 
 module.exports = {
-    parseFilter
+    parseFilter,
+    DB
 };