--- a/test/test_utils/index.js
+++ b/test/test_utils/index.js
@@ -1,0 +1,5 @@
+var parseFilter = require('./parse_filter');
+
+module.exports = {
+    parseFilter
+};