The benchmarks can be run using `npm run-script -s perf`.