@bkutil 2016-02-28
After rewriting the `toInflux` serialisation, the throughput improved nicely:

before:
```
serializer.toInflux x 33,620 ops/sec ±3.14% (84 runs sampled)
AdapterWrite.serialize, 1k points/call x 30.77 ops/sec ±3.93% (50 runs sampled)
AdapterWrite.write, stubbed request, 1k points/call x 28.59 ops/sec ±5.87% (65 runs sampled)
AdapterWrite.write, live request, 1k points/call x 19.88 ops/sec ±7.36% (49 runs sampled)
```

after:
```
serializer.toInflux x 152,627 ops/sec ±0.69% (92 runs sampled)
AdapterWrite.serialize, 1k points/call x 129 ops/sec ±1.01% (71 runs sampled)
AdapterWrite.write, stubbed request, 1k points/call x 128 ops/sec ±2.00% (75 runs sampled)
AdapterWrite.write, live request, 1k points/call x 50.01 ops/sec ±7.86% (57 runs sampled)
```