@rlgomes 2016-03-04
Quick glance this looks great, a few suggestions come to mind:

 1. Why not make the performance tests mocha tests ? you can use `mocha-logger` to output the performance numbers instead of console.log and at least running things is a bit "simpler" by doing `gulp performance` (of course we need to up the timeouts to handle the fact these run for longer than a regular mocha test). 
 2. wonder if we can also start saving the results of each run to a "peformance/results/v0.5.txt" where we rerun the tests in a somewhat controlled environment and record the results there just for tracking ? Depending on how long these take maybe our controlled environment would be to run and save the output from Travis itself ? (we can discuss this offline)
 3. We should definitely apply this on our other adapters so we can give people a pretty good idea of what to expect when read/writing to a given adapter. I would say lets open issues on each of the existing adapters if others are in agreement that this would be a nice thing to have in each adapter and now that I'm writing this we should also do the same for the juttle engine itself and that way we'd know how performant the base engine is at doing some simple point pushing through the pipeline.

None of the above need to block this but I do think we need to figure out how we go about rerunning these in a way that can help us identify significant performance regressions before we'd ship something.